request with several parameters

+++ delList.php
@@ -4,8 +4,9 @@ require_once('Connector.php');
 $connector = new Connector();
 $link = $connector->connect();
 
-$userId = $_SESSION['id'];
+if (!empty($_SESSION)){
+    $userId = $_SESSION['id'];
 
-$link->query("DELETE FROM tasks WHERE user_id = '$userId'");
-
+    $link->query("DELETE FROM tasks WHERE user_id = '$userId'");
+}
 header('Location: /tasklist.php');
